+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "gdsc.stydyjams.newsapp"
+        applicationId "gdsc.stydyjams.newapp"
         minSdk 21
         targetSdk 31
         versionCode 1
