+++ app/src/androidTest/java/gdsc/stydyjams/newapp/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package gdsc.stydyjams.newsapp
+package gdsc.stydyjams.newapp
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("gdsc.stydyjams.newsapp", appContext.packageName)
+        assertEquals("gdsc.stydyjams.newapp", appContext.packageName)
     }
 }
