+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="gdsc.stydyjams.newsapp">
+    package="gdsc.stydyjams.newapp">
 
     <application
         android:allowBackup="true"
@@ -8,10 +8,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.NewsApp">
-        <activity
-            android:name=".RegisterActivity"
-            android:exported="false" />
+        android:theme="@style/Theme.NewApp">
         <activity
             android:name=".LoginActivity"
             android:exported="false" />
