+++ app/src/main/java/gdsc/stydyjams/newapp/MainActivity.kt
@@ -1,23 +1,21 @@
-package gdsc.stydyjams.newsapp
+package gdsc.stydyjams.newapp
 
 import android.content.Intent
-import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        // check if user is logged in
-        val loggedIn = false
-        if (!loggedIn) {
-            // send to login page
+        var loggedIn = false
+        //dashboard
+        if(!loggedIn)
+        {
+            //redirect to the login page
             val loginIntent = Intent(this, LoginActivity::class.java)
             startActivity(loginIntent)
-            this.finish()
         }
-        // else stay here
-
     }
 }
